pyproject.toml
[project]
name = "email-waitlist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-pika>=9.5.8",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.123.7",
    "httpx>=0.28.1",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
